Don't bother checking line count, try to squeeze everything
Manipulating the DOM to determine element height is a clever idea, but
it's a premature optimization:

(1) it's additional complexity,

(2) the 10 pixel fudge factor leaves its behavior on small text
questionable,

(3) this implementation doesn't work when a div's html starts with tags
instead of text, because the first split() word ends up nonvisible so
firstWord.offsetHeight is zero, and

(4) it's faster to simply try squeezing and see what happens. In my
rudimentary testing, headlineIsMultipleLines ran in 2.66ms, but simply
running squeezeContainer completed in 2.73ms, so it's making things
slower. My numbers may not be super accurate, but they would have to be
very far off for my conclusion to be backwards.